dev-python/yt-dlp-ejs: new package, add 0.1.0
This has very little python, its purpose is to provide the .js
files generated by deno and then the minimal python code just
feeds these files as a string to yt-dlp.

Given tests are not wired (may never be, they use deno and have not
managed to get them to run at all yet), and that this is allarches,
went ahead and self-copied yt-dlp keywords. Note would be meaningless
to run tests for PYTHON_COMPAT given they do not use python at all.

yt-dlp currently only allows specific sha512sum for the .js files
meaning that this will need to be kept pinned with yt-dlp versions.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>