net-misc/openssh: backport new key exchange (and other fixes) to 9.8_p1, 9.9_p2
Upstream asked that a new key exchange algorithm be backported to older
releases being shipped by distros [0], so let's do that. While at it,
include other changes from the corresponding release branches.

As for why we're still packaging older releases anyway: it's because
older OpenSSH is useful to connect to old machines, and for security
testing.

[0] https://marc.info/?l=openssh-unix-dev&m=175495631413568&w=2

Signed-off-by: Sam James <sam@gentoo.org>