dev-libs/protobuf: filter LTO in 31.1
A bunch of people have hit this now that 31.1 has been stabilised so
let's filter LTO here for now. What is not clear is why it only happens
in some cases, e.g. I can't hit it on my workstation, but I can in
a stable chroot.

My guess is it uses some installed headers if available (and may require
tests not be enabled as well).

Bug: https://bugs.gentoo.org/963340
Signed-off-by: Sam James <sam@gentoo.org>