app-shells/bash: overhaul SLOT 0 ebuilds; minimise their differences
The bash ebuilds in SLOT 0 continue to diverge from one another in small
but irksome ways, which makes working on them a chore. As such, the main
purpose of this commit is to minimise the differences between them.
Beyond that, several other changes have been made. All of the changes
are described herewith.

As regards the matter of difference minimisation:

- order the patterns of the first case statement in a consistent way
- declare BASH_COMMIT within the scope of the first case statement
- always declare BASH_COMMIT (even if as a null string or dead code)
- remove a redundant ;; operator from the first case statement
- always acknowledge that BASH_COMMIT might be set
- declare S within the scope of the second case statement
- consistently mention that Chet advises against the use of byacc

In pkg_postinst(), incorporate a warning to the effect of the
"15-gentoo-bashrc-check.bash" file being safe to delete.

Restructure the pkg_postinst() code so as to make it easier to
incorporate additional, version-dependent, upgrade notices.

In pkg_postinst(), if the terminal is found to be more than 80 columns
wide, invoke the fmt(1) utility so as to reflow the upgrade notice(s).

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>