app-shells/bash: revbump 5.2, 5.3 and 5.4 for improved title setting
This commit introduces the following versions.

- 5.2_p37-r5
- 5.3_p3-r3
- bash-5.4_alpha_pre20250918-r2 (replacing -r1)

They improve the window title setting behaviour by using the ${param@P}
form of expansion to expand "\u@\h \W" as a prompt string. A notable
consequence of this is that, upon switching to the home directory, the
current working directory will be shown as the <tilde> character again.

The value of PROMPT_DIRTRIM is now respected. If this variable is unset,
the use of the \W prompt string escape will prevail, with the current
working directory typically being shown as its basename. If set to 0 or
greater, \w will be used instead, which may be trimmed. This means that
the title can be made to show the full path by setting PROMPT_DIRTRIM=0.

Also, whitelist xterm-ghostty for XTWINOPS support.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>