media-video/ffmpeg: remove IUSE=npp
ffmpeg has deprecated[1] libnpp support, seems does not intend to fix
it for cuda 13.0, and has added a ./configure check to abort if it
finds >=13.0 (in live). Rather than update our npp cuda 13 patch
to bypass that check or add upper bounds, let's just drop support.

Debated making that change only in the live ebuild unless it gets
backported (might), but given cuda 13 breaks npp for older nvidia
cards at /runtime/ that may lead users to report bugs to ffmpeg
(not us) when it's not even supposed to be buildable with it, feel
we're better off killing it.

USE=libplacebo and/or USE=cuda should be sufficient alternatives.
Users that really want to keep using npp while it still works can
use EXTRA_ECONF=--enable-libnpp but no support will be given.

Will revisit if upstream ffmpeg un-deprecates it and handles issues.

[1] https://code.ffmpeg.org/FFmpeg/FFmpeg/commit/994a3684

Closes: https://bugs.gentoo.org/695210 (obsolete)
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>