dev-qt/qt-docs: add comment to explain unpacker depends usage
It's kind of a unintended way to use it by passing a non-existing
archive, and may also seem useless from an onlooker when it's to
save from looping through 155 SRC_URI archives in global scope just
to find that at least one of them has a .7z.

...and qt6_docs_generate_metadata is already busy enough in global
scope in this ebuild to not need this big loop on top.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>