net-misc/openssh: wire up more tests
Note that the 'tests' target does run them all, but in most cases, we don't
use that because of the shell needed for the 'portage' user.

So, wire up the following:

* 'unit', added in e7429f2be8643e1100380a8a7389d85cc286c8fe upstream (2014);
* 'file-tests', added in 159e987a54d92ccd73875e7581ffc64e8927a715 upstream (2019).

It would be nice to perhaps wire up more of the interop tests w/ dropbear
and twisted (IIRC) at some point.

Signed-off-by: Sam James <sam@gentoo.org>