dev-db/mariadb-connector-c: add 3.3.17
I spent some time understanding why the QA Notice for
CLIENT_PLUGIN_AUTH_GSSAPI_CLIENT pops up, and I found out that this
variable is only available if the GSSAPI dependency requirements are
met. It is unrelated to AUTH_GSSAPI_PLUGIN_TYPE, which is basically
useless these days.

I have added a comment to clarify that the QA Notice is falsely positive
above the related line.

Signed-off-by: Petr Vaněk <arkamar@gentoo.org>