sys-libs/libapparmor: complete wiring up tests for Python
Do a nasty hack to avoid running the Python tests in the first/general
build, otherwise it'd run for the wrong impl and may not find the built
module if not already installed.

Still not super happy with all of this and would prefer that we had clean
separation of the non-Python and then Python builds but for another day.

Bug: https://bugs.gentoo.org/623550
Bug: https://bugs.gentoo.org/625448
Bug: https://bugs.gentoo.org/764779
Signed-off-by: Sam James <sam@gentoo.org>