dev-qt/qtbase: skip flaky tst_wayland_cursor in 6.10
Randomly had this fail then pass, test sounds timing related.
It's a small test case and probably not super important to check
here, so not going to worry about it. Unsure if even relevant
without a real wayland compositor and a cursor (offscreen).

For 6.9, this test was in dev-qt/qtwayland instead, albeit I
never had it fail there (yet) so will leave that alone.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>