sys-apps/isapnptools: fix compile on musl
Copied this patch from ::musl overlay (whose origin is unknown
since it appeared as part of initial commit in that repo).

The change itself consists of 2 fixes:

1. Adding "#include <stdlib.h>" which is safe
2. converting conditional include into consistent "#include <sys/io.h>"
   which should work on all libc we care about in Gentoo.

I don't see a reason to cause rebuild with this patch, since where it
managed to build work as needed, so no need for revbump.

Closes: https://bugs.gentoo.org/712764
Closes: https://bugs.gentoo.org/889557
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>