cmake.eclass: disable automagic LTO
Pass -DCMAKE_INTERPROCEDURAL_OPTIMIZATION=OFF to avoid automagically
enabling LTO if the compiler supports it.

Note that unlike what we do w/ meson.eclass, we can't do some magic
with `tc-is-lto` here because CMake doesn't allow refining the parallelism
used. But that's fine.

This change still allows using LTO via *FLAGS which is what we tell
people to do.

Closes: https://bugs.gentoo.org/927157
Signed-off-by: Sam James <sam@gentoo.org>
Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>