dev-libs/libuv: disable io-uring for ppc32
Works around a crash with CMake (-> libuv -> io-uring) on ppc32.

There are reports of this working with libuv-1.51.0 and cmake-4.1.2 but
the fix was likely on the kernel side which means it's not something we
can rely on, especially for an arch like ppc32.

This could be submitted upstream if we can identify what actually fixed
it and therefore the first good kernel version to check for, as we do
for hppa.

Bug: https://bugs.gentoo.org/941738
Signed-off-by: Sam James <sam@gentoo.org>