dev-qt/qtdeclarative: backport fix for QTBUG-140018
Sounds that it won't make it in 6.10.0, will be fixed in 6.10.1.

Plasma did a workaround for this but think(?) the workaround is not
in-tree yet, may either way affect other things.

Stable 6.9.2 is also affected but leaving it alone given the size of
the patch to get some testing first esp. given the patch is meant
for 6.10.0 and not 6.9.x (6.9 branch is closed).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>