sys-fs/reiserfsprogs: fix building on musl
The patch addition is safe for glibc, since all additions are behind
"#if !defined(__GLIBC__)". The patches are from void-linux.

On musl it also requires sys-libs/obstack-standalone, and manual
add of "-lobstack" to linking (needed by debugreiserfs).

Also noted that we are missing sys-fs/e2fsprogs in R/DEP, for which
I've revbumped.

Closes: https://bugs.gentoo.org/832972
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>