sys-kernel/linux-headers: Add rsync to BDEPEND
This program is usually *always* available in gentoo, being a direct
dependency of portage, and some other system packages. Still, it's
useful to list dependencies like this in bootstrap scenarios.

It is my belief that this dependency was introduced in 5.3, with the
following commit:
https://github.com/torvalds/linux/commit/59b2bd05f5f4dc62979c2e82ddd384f07e8f10bc

Closes: https://bugs.gentoo.org/963902
Signed-off-by: Esteve Varela Colominas <esteve.varela@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/44088
Closes: https://github.com/gentoo/gentoo/pull/44088
Signed-off-by: Sam James <sam@gentoo.org>