dev-cpp/rapidyaml: add 0.9.0, 0.10.0
This been un-last-rited for upcoming kde-plasma/union (currently only
in kde overlay), but union is currently broken with both 0.10.0 and
old stable 0.6.0 so also adding 0.9.0 for the interim.

This is a simplified version of the original ebuild to make maintenance
easier (do not really want to spend more time than necessary on this
package anymore, feel free to take over maintenance if want to give
more care to this package).

Afaik union is likely still far from release, so could potentially be
last rited again if final version doesn't need it anymore.

Bug: https://bugs.gentoo.org/963014
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>