x11-wm/mutter: add 49.0
Depends on glycin-2.0 (#44092). There are various upstream patches that
should be applied here but since 49.1 is releasing today (or tomorrow
(or the day after that)) I am going to exclude them for now so we don't
have to juggle patches (unless someone desparately needs 49.0 and not 49.1).

A new dependency on devkit is also introduced for screensharing and nesting.

Signed-off-by: Swagtoy <me@ow.swag.toys>
Signed-off-by: Lukas Schmelting <lschmelting@posteo.com>
Co-authored-by: Lukas Schmelting <lschmelting@posteo.com>