gui-libs/egl-wayland2: new package, add 1.0.0_rc
Meant to replace gui-libs/egl-wayland, but should still be considered
experimental (use as your own risks). Requires a modern wayland
compositor with dma-buf support and >=nvidia-drivers-560.

Adding for testing given NVIDIA sometimes asks users to try it in case
it fixes issues.

egl-wayland and egl-wayland2 can be installed at same time, in which
case the latter will be used if >=nvidia-drivers-560

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>