dev-util/hip: add 7.0.2
This commit includes hip-7.0.2-fix-libcxx-noinline.patch, which solves
an old issue with libc++. At some point of time this problem should be
reported to upstream, but at this point it is unclear who should fix it
(llvm-project or ROCm), so there is no report. This patch allows to use
rocm_use_clang function from rocm.eclass instead of rocm_use_hipcc.
ROCm gradually deprecates hipcc for simplification and due to
side-effects. Migration to clang was done just recently in the
https://github.com/ROCm/TheRock repository; now Gentoo uses clang as
well (mostly).

Bug: https://bugs.gentoo.org/964799

Signed-off-by: Sv. Lockal <lockalsash@gmail.com>