dev-python/libvirt-python: Reflect tar name change in live ebuild
As of [1] libvirt-python release tarballs are named
'libvirt_python'. Reflect this name change in the live ebuild,
since ebuild for new releases are going to be copied from it.

This also removes replacement of "_rc" to "-rc" which might have
been useful back in the day when python bindings lived in libvirt
main repository, but they were moved to a separate repo ~12 years
ago [2].

1: https://gitlab.com/libvirt/libvirt-python/-/commit/a2047351b2889f611e229725971fd8cf7b3d284c
2: https://gitlab.com/libvirt/libvirt/-/commit/a7a1244a739d134a9aca7f76c12d63043810c077
Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>