sys-apps/locale-gen: Run mkconfig only where SUPPORTED can be found
The mkconfig utility is used to incorporate commented examples of every
supported glibc locale into the /etc/locale.gen file. However, there are
sitautions in which the SUPPORTED file will not be present, such as
where cross-compiling from a musl environment.

For now, address this issue by only running mkconfig in the case that
the SUPPORTED file is found to exist.

Closes: https://bugs.gentoo.org/964382
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>