dev-util/pwntools: avoid using blocker for broken dep
Portage doesn't backtrack around blockers like this, (weak) blockers are purely
for file collisions (or in exceptional cases, strong blockers where the
existence of one package totally breaks another -- but they don't work
as "forward dependencies").

Just use a < dep for now. When a fixed version of unicorn is available,
we can swap to || ( ... ).

Fixes: c05064e5a16cf71ced2e20b31b04ba3d7ce33db5
Signed-off-by: Sam James <sam@gentoo.org>