gui-libs/kddockwidgets: set CMAKE_QA_COMPAT_SKIP=1
The ECM code paths are not currently used here (thus why it's also
missing from BDEPEND), upstream may need to care for this in the
future but it should not affect us unless that changes.

As for cmake 4 itself, min ver is 3.15 which should be of no concern
for the foreseeable future.

Closes: https://bugs.gentoo.org/964536
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>