dev-libs/rocm-core: new package, add 7.0.2
This is a small package which includes header and verson library.
Previously the same header was added by dev-util/hip ebuild,
but other packages like amdsmi also use rocm-core library, so it is
better to install them in a separate packages (as intended by upstream).

Bug: https://bugs.gentoo.org/964799

Signed-off-by: Sv. Lockal <lockalsash@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/44165
Signed-off-by: Sam James <sam@gentoo.org>