app-emulation/wine-staging: drop 10.10, 10.13, 10.14
wrt 10.10 this was the last version to support ESYNC and was kept
for longer because of it, but in theory this shouldn't be needed
anymore now that NTSYNC should be usable (haven't really tested
that myself though).

If really needed, 10.0 is still there for now, or wine-proton supports
ESYNC+FSYNC (but makes more sense to use FSYNC there).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>