media-video/vlc: add <ffmpeg-8 upper bound for now
Not quite sure for the state of the upstream fix given it seems
undecided/confused about how to handle it for old versions and I
haven't reviewed it myself.

It'd likely be easier to do our own simple fix and depend on >=6.1,
but I'll just throw a upper bound for now given about to rekeyword
ffmpeg-8 and want to avoid failures. If someone else wants to look
at this, by all means please do, hopefully this upper bound should
be short lived.

(note stable version has an effective upper bound due to [postproc])

Bug: https://bugs.gentoo.org/961436
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>