net-misc/yt-dlp: add 2025.10.22_pre w/ WIP deno usage (unkeyworded)
For early testing esp. given a lot of users are impatient about this.
You can verify that deno is being used by running yt-dlp with `-v`:

    [debug] [youtube] [jsc:deno] Running deno: deno [...]

Be warned that this is still WIP upstream (PR not merged and marked as
Draft) with a lot of TODOs including for documentation, so use at your
own risks. It may also not necessarily magically fix issues if youtube
isn't working with normal releases for you, so keep your expectations
low.

Tarball is generated with a local script from what the PR is currently
based on and may be lacking some commits that are older than than this
version (aka it is not rebased).

Likely will not update this regularly unless there is a specific reason
to, so it may fall behind the PR until a proper release. Also no plans
to make this work with a live ebuild meanwhile.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>