rust.eclass: introduce RUST_NEEDS_LLVM_BACKEND instead
Having RUST_SUPPORTED_CODEGEN_BACKENDS is a bit awkward to handle for
dependencies because you'd need to do:
```
	|| (
		dev-lang/rust:1.90.0[rust_codegen_backend_llvm]
		dev-lang/rust-bin:1.90.0[rust_codegen_backend_llvm]
		dev-lang/rust:1.90.0[rust_codegen_backend_gcc]
		dev-lang/rust-bin:1.90.0[rust_codegen_backend_gcc]
	)
```

and so on. An earlier attempt already got this wrong and tried to do:
```
	|| (
		dev-lang/rust:1.90.0[rust_codegen_backend_llvm,rust_codegen_backend_gcc]
		...
	)
```

Let's avoid the complexity both in the eclass and the generated dependenceis
if possible. Certainly not worth it at at this early stage.

Define RUST_NEEDS_LLVM_BACKEND if the package needs the LLVM codegen
backend (which is the default) but doesn't need LLVM otherwise (and hence
llvm-r1.eclass). If it is set, we handle that in pkg_setup which should
be enough for now as the LLVM backend is forced on.

My original proposal for posterity, as we may want to reuse something
like it in future:
> Add RUST_SUPPORTED_CODEGEN_BACKENDS for ebuilds to declare which codegen
> backends they support. By default, this is the same as RUST_CODEGEN_BACKENDS
> (so 'llvm' and 'gcc'). ebuilds can set RUST_SUPPORTED_CODEGEN_BACKENDS=( llvm )
> to workaround bugs with alternative backends.
>
> As future work, we can modify RUSTFLAGS to pass -Zcodegen-backend=PATH or
> perhaps introduce another USE_EXPAND (bleh) to allow users to control the
> default to determine when we need to pass that (the first codegen backend
> specified when building Rust is the default). This is complicated by the fact
> I'm not sure all Rust packages really respect RUSTFLAGS fully, see e.g.
> bug #959723.
>
> The priority at this time is to avoid disruption and facilitate experimenting.
>
> TL;DR:
> * If you want to use the GCC backend, either disable the LLVM one too by
>   setting RUST_CODEGEN_BACKENDS, or pass -Zcodegen-backend=PATH in RUSTFLAGS.
> * I hope we can make rust.eclass handle this in some more advanced way
>  in future.

Bug: https://bugs.gentoo.org/959723
Bug: https://bugs.gentoo.org/965182
Signed-off-by: Sam James <sam@gentoo.org>