app-alternatives/gpg: depend on fixed gpg
Depend on a version of (stable) gnupg which always provides fallback
symlinks. This doesn't cover the ~arch case but it matters less as there's
no official binpkgs for that, so less likely something goes wrong, and
the old ebuilds are gone already, so it's a narrow enough window I think.

See also 133742ca3a9497606262fab40ed6f4f965782a7c and
133742ca3a9497606262fab40ed6f4f965782a7c.

Signed-off-by: Sam James <sam@gentoo.org>