sci-libs/rocBLAS: switch to sci-libs/flexiblas and fix dependencies
sci-libs/flexiblas seems to be a decent solution given that it cblas
is only used for validation.

Flag video_cards_amdgpu was dropped, because it was non-switchable;
we require to select AMD GPU with REQUIRED_USE; attempt to disable
video_cards_amdgpu after that leads to compilation error.

Yes, technically it is possible to build rocBLAS for CUDA, but there is
no request for this at this moment.

Closes: https://bugs.gentoo.org/965167
Signed-off-by: Sv. Lockal <lockalsash@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/44373
Signed-off-by: Sam James <sam@gentoo.org>