dev-perl/File-Slurper: add 0.14.0-r1 with USE=minimal support
If available, File::Slurper loads the PerlIO::utf8_strict module for
faster UTF-8 coding with better diagnostics. However, File::Slurper is
designed to work perfectly well in the case that only the (core) Encode
module is available.

This revision makes it possible to forgo PerlIO::utf8_strict by enabling
the "minimal" USE flag. The reason for not choosing "xs" as the flag
name is that File::Slurper implements no XS code of its own.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>