sys-cluster/ceph: remove libxml2 deps
If you search for libxml in the source code, you'll see that it only finds .md,
.sh, .control, .in, .dockerfile, .txt, and .json files. But no code files,
because I stopped when updating libxml2 on a server without compiling Ceph.

It didn't show any broken libraries for Ceph, and that's where I searched with
revdep-rebuild, and it didn't show any problems. Therefore, I see that
requiring libxml2 in Ceph is unnecessary.

[sam: Wrap commit message.]
[sam: This seems to go back to 45ceab32803179e38d88eb875ead089b3926274c in
pre-git days.]

Signed-off-by: Fco Javier Felix <ffelix@inode64.com>
Part-of: https://github.com/gentoo/gentoo/pull/44413
Closes: https://github.com/gentoo/gentoo/pull/44413
Signed-off-by: Sam James <sam@gentoo.org>