www-client/epiphany: add 49.1
This also introduces some patches to disable criticals. One of these is
on the Webkit process itself, which even upstream Webkit disapproves
of. Webkit usually handles GStreamer issues themselves, but as a result
GStreamer ends up throwing a critical in some cases (even if the error
was already handled) which causes the Webkit process to die, thanks to Epiphany.

In general, upstream Epiphany needs better error handling procedures
instead of just dying, like a toast or a popup for crititcals. Upstream
claims that this is done to "encourage users to report any bugs," but
instead it breaks even working on websites. I assume Gentoo users would
know how to find and report a bug if something seems off inside of
Epiphany, so I don't think this is neccesary.

As I type this, Github.com and bitbucket do not work due to an issue in
GStreamer, though I have asked parona to backport patches for that
already anyway. These patches fix this as well.

FWIW, Epiphany is the only program to do this practice, it is not
encouraged at all by upstream, nor even some people from WebKit like
this.

https://github.com/WebKit/WebKit/pull/28651
https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1488

Signed-off-by: Swagtoy <me@ow.swag.toys>