dev-ruby/rdoc: fix compilation
rubocop has been introduced and made optional, but not for the generate
task that we use to regenerate the files. Patch out the rubocop step
since rubocop is not packaged in ::gentoo and since we can't guarantee a
working version. This will mean that our generated files will not be
identical to the upstream gem, but the changes should be cosmetic only.

Closes: https://bugs.gentoo.org/965427
Signed-off-by: Hans de Graaff <graaff@gentoo.org>