nginx-module.eclass: Add NGINX_MOD_CONFIG_DIR, fixup www-nginx/ngx-naxsi
This might be useful for packages having "config" in a non-root
directory. For example, imagine the following directory tree for some
NGINX module:

+--nginx/
|  +--config
|  +--ngx_module.c
|
+--lib/
|  +--defective.c
|
+--README.md

In case defective.c needs to be patched, we need to know both the path
to the config file and the root directory of the module. Without this
patch, we only know the path to config, making it very hard to patch
defective.c

This commit introduces a new variable NGINX_MOD_CONFIG_DIR specifying
the subdirectory of NGINX_MOD_S containing the config script.
NGINX_MOD_S semantics change to "the root directory of the module
sources".

www-nginx/ngx-naxsi was a special case because it is the only module
actually having the config script in a subdirectory of NGINX_MOD_S. This
commit edits its patch and ebuild accordingly.

Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>