x11-drivers/nvidia-drivers: add 580.105.08
aarch64 files are missing for some reason, so drop keyword for now
and will restore once I see them.

Could've waited but users tend to be impatient if do not bump this
right away.

On a side-note, noticed there may be some regression with (some) custom
higher refresh Modelines on Xorg albeit that is seldom used and normal
modes should function as usual (my case is special because my monitor's
edid does not report the 75fps mode that it's supposed to be able to
use and in fact can, and been using a Modeline to workaround that --
this version takes me back to 60fps and haven't figured a workaround
yet).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>