gnome-base/gnome-session: exclude gnome-shell from session query
Or else on systemd the user can't logout of the session, on OpenRC they
technically can (all the apps will get killed except gnome-shell) but
gnome-shell will show up as an unresponsive program and the session will
fail to end i.e the user can't logout. Only needed with the backported
patch for the bug with <gnome-shell-49.

Bug: https://bugs.gentoo.org/959295
Signed-off-by: zyxhere💭 <zyx@envs.net>