app-editors/vim-core: provide an escape hatch to avoid the system vimrc
It contains some problems or otherwise unnecessary settings for users
with highly customized personal configurations.

Here's taste of what it takes to disable the system vimrc's effects
otherwise:

    " Undo /etc/vim/vimrc settings first
    set backspace& autoindent& history& ruler& viminfo& suffixes& numberwidth&
    set modeline& fileencodings& guifontset& hlsearch&
    silent! unmap Q
    " - Skip turning off syntax, since my vimrc is smart enough to already check
    "   t_Co, has('gui_running'), _and_ !exists("g:syntax_on") before calling
    "   `syntax enabled`.
    " - Skip undoing term settings because they don't apply to me
    " - Ignore filetype plugin on, even though it comes way too early; I do it again
    "   later.
    " - &shell should not be empty for me
    " - I set is_bash set later, fine
    " Hmph. These autocommand set options like tabstop, shiftwidth, expandtab
    " globally insetad of locally. Broken.
    silent! autocmd! gentoo
    " Don't need this variable: I have a vimrc
    unlet! g:skip_defaults_vim
    " Can't revert the Filetype * autocommand for omnifunc because it's not in a
    " group, so I'll just have to live with duplicates.

Obviously this doesn't cover all the effects, some of which are
next-to-impossible to reliably revert, and some of the changes or
omissions are relevant only to certain users. But for folks who just
want a blank slate, we need to provide an escape hatch. Unlike
defaults.vim, we cannot provide a simple mechanism to skip the whole
file: by the time we would set the variable, in a local or user vimrc,
it's too late!

Instead, we first source any files in runtimepath called
"skip_etc_vimrc.vim": they have an opportunity to go early and perform
tasks like

    let skip_etc_vimrc = 1

to skip the remainder of the script. It's a complicated workaround, but
it avoids a new USE flag.

Signed-off-by: D. Ben Knoble <ben.knoble+github@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/44504
Signed-off-by: Sam James <sam@gentoo.org>