gnome-extra/cinnamon: fix runtime w/o networkmanager
In versions 6.4.10 and 6.4.13, the build option disable_networkmanager is
broken and does not work.

When disable_networkmanager is set to true, the build succeeds, but it
will not start with startx.

In the upstream master branch, the disable_networkmanager option does not exist;
instead, there is a new option called nm_agent. This has not been released as
of 2025 November 6.

Closes: https://bugs.gentoo.org/960231
Signed-off-by: osakafinance <deceiver_unbundle593@sl.projectsegfau.lt>
Part-of: https://github.com/gentoo/gentoo/pull/44513
Closes: https://github.com/gentoo/gentoo/pull/44513
Signed-off-by: Sam James <sam@gentoo.org>