sys-devel/binutils: don't try -static-pie at all in the sanity check
rcrt1.o isn't always available and static-pie support can be buggy (which
is why it's interesting to test, but if we can't rely on it being available,
we also can't rely on it at least without some checking being OK in some
glibc version too). Need to try again with a has_version check for glibc[static-pie]
at least and possibly more.

Closes: https://bugs.gentoo.org/965478
Signed-off-by: Sam James <sam@gentoo.org>