media-video/ffmpeg: skip tests that are picky about zlib for zlib-ng
At same time setup a placeholder variable to make it easy to skip
more tests in the future if needed given need to be all be gathered
in a single configure option. Will likely keep that there even no
longer need to skip anything.

A downside of --ignore-tests is that configure options are recorded
and users will see the whole list of test skips in ffmpeg invocations
(which will vary depending on which zlib it built against, not that
it actually changes the code).

Haven't verified the skip list for old ffmpeg versions, so it may
be skipping some not-yet-existing tests (does pass though).

Closes: https://bugs.gentoo.org/965737
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>