ninja-utils.eclass: Respect NINJA for NINJA_DEPEND again
Make NINJA_DEPEND respect the value of NINJA again.  This was changed
while updating for app-alternatives/ninja.  However, it is still valid
to override NINJA in ebuilds, in which case we want NINJA_DEPEND
to refer to the appropriate implementation rather than any alternative.

Fixes: 2fa125d055dcf ("ninja-utils.eclass: use app-alternatives/ninja")
Bug: https://bugs.gentoo.org/922639
Signed-off-by: Michał Górny <mgorny@gentoo.org>
Part-of: https://github.com/gentoo/gentoo/pull/44539
Signed-off-by: Sam James <sam@gentoo.org>