dev-java/log4j-core: simplify handling of the '-processor' option
This package generates coding by using annotation processors. These
annotation processors and their '-processorpath' should be passed to
ejavac via the JAVAC_ARGS eclass variable. To do so, the processors must
either exist in another package or be compiled first. For compilation,
it needs '-classpath' which is built by java-pkg-simple_getclasspath().
This function however is marked '@INTERNAL' and cannot be used in an
ebuild. As a workaround, we run the whole java-pkg-simple_src_compile()
twice, first time for getting the processor as a class and second time
for letting the processor process the annotations during the regular
compilation.

Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>