net-im/signal-desktop-bin: fix app icon on wayland
The wayland appid and the basename .desktop do not match, therefore kde
(and likely others) can not find a .desktop file for windows and will
assign them a default wayland icon instead of the one from the desktop
file.
It is a problem well known upstream with no fix currently in sight.

Signed-off-by: Henning Schild <henning@hennsch.de>