media-gfx/imv: add 5.0.0
Was starting to consider dropping this to maintainer-needed and
expecting eventual last-rites, but seems it got a new maintainer
upstream.

freeimage support was dropped which IMO is good given freeimage is in a
garbage state and hoping we can remove it from Gentoo eventually. This
does lose support for some obscure formats, but it keeps bmp+webp by
newly using libnsbmp+libwebp (personally I would recommend libvips to
upstream if want wider image support, but there is many alternatives).
As a bonus, removing freeimage obsoletes the old (removed) nsgif patch
to ensure it's used over freeimage so gif can be animated.

Also adds support for qoi and farbfeld, but only enabling the latter
for now given qoi is unpackaged and I do not intend to maintain it
(it is just a tiny header-only package though). Not exposing farbfeld
as a USE given support is builtin and VERY small (easy implementation
is a design goal of the format), not that many users will need this.

Also re-add USE=gif given this can use libnsgif-1.0.0 now, albeit the
upstream fix used a old patch that I had formerly rejected downstream
because the meson.build checks for == 1.0.0 rather than >= 1.0.0 and
will break when libnsgif-1.0.1 or so releases even if it does not break
compatibility. Quickfix for now but still needs to be reported.

Given freeimage was default, at least enable jpeg+png by default to not
be near-useless on non-desktop profiles. Given people using bmp is rare
nowadays, will let users enable that one manually if needed though.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>