dev-util/mingw64-toolchain: add 13.0.0_p1 (gcc-15.2+binutils-2.45.1)
Kind of overdue for some fixes in >=binutils-2.45 that were requiring
patches in older Wine versions with 2.44, but was kept on hold due to
needing further fixes in 2.45.1 (thanks to sam for the release, so do
not need to use a snapshot).

As far as I'm aware there is no specific reason to use a gcc snapshot
for our use case, so 15.2.0 should be fine like 15.1.0 still is afaik.

Believe no ebuild changes are needed, but have not thoroughly inspected
differences between 15.1->15.2 and 2.44->2.45.1.

Briefly tried both the oldest and newest Wine versions we have in-tree
(not all) but, like usual, hard to say if this regress anything at
runtime.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>