sys-devel/distcc: do not rename gui binary anymore
While the renaming seems convenient it can cause issues in the whole xdg
system, where binary names desktop file names etc. have to match for
everything to work as expected.

The rename was incomplete and a complete rename involves too many
things, so drop it and do what upstream does.

Closes: https://bugs.gentoo.org/965994
Signed-off-by: Henning Schild <henning@hennsch.de>