net-im/signal-desktop-bin: add 7.79.0
The wayland appid and the basename .desktop do not match, therefore kde
(and likely others) can not find a .desktop file for windows and will
assign them a default wayland icon instead of the one from the desktop
file.
It is a problem well known upstream with no fix currently in sight.

Co-authored-by: Henning Schild <henning@hennsch.de>
Closes: https://github.com/gentoo/gentoo/pull/44602
Signed-off-by: Robert Siebeck <gentoo.2019@r123.de>