Revert "games-strategy/endless-sky: upgrade to 0.10.16"
This reverts commit 1e8b70cbcafa53ced9726ba09d3b2c5ed9882f09.

This did a git mv, and it wouldn't hurt to keep the old version
for a bit in case of issues.

...and seems commit author did not check for new dependencies (again).
Missing avif wrt bug #966057 and also flac[cxx]. At a glance, these
seem to be the only issues and it seems to run and test fine though.

Will re-add 0.10.16 in a follow up commit with a revbump.

Bug: https://bugs.gentoo.org/966057
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>